Adds measure pass to the FilterSystem.
This is the initial commit for the measuring pass API to PIXI.FilterSystem; it will allow filters to specify the area required by them to render into the output.

Use Cases:
When a object is partially visible, autoFit=false or true might not be the optimal solution. The filter doesn't need the whole object, neither can it work by just the visible portion (e.g. blur).

Backward Compatibility:
> measure(): A default onMeasure() is provided by the Filter class. It will cause the filter ask the input to be the whole object (if autoFit=false) or just the specified output (if autoFit=true).

> Vertex shader: inputRect & outputRect are two new uniforms. The vertex shader will automatically map the texture coords onto the input texture.